Add pytest, re-wire ui connection endpoint
