+++ airflow/api_fastapi/core_api/openapi/v1-generated.yaml
@@ -58,6 +58,26 @@ paths:
                 $ref: '#/components/schemas/HTTPExceptionResponse'
       security:
       - OAuth2PasswordBearer: []
+  /ui/connections/hook_meta:
+    get:
+      tags:
+      - Connection
+      summary: Hook Meta Data
+      description: Retrieve information about available connection types (hook classes)
+        and their parameters.
+      operationId: hook_meta_data
+      responses:
+        '200':
+          description: Successful Response
+          content:
+            application/json:
+              schema:
+                items:
+                  $ref: '#/components/schemas/ConnectionHookMetaData'
+                type: array
+                title: Response Hook Meta Data
+      security:
+      - OAuth2PasswordBearer: []
   /ui/dags/recent_dag_runs:
     get:
       tags:
@@ -8121,6 +8141,51 @@ components:
       - total_entries
       title: ConnectionCollectionResponse
       description: Connection Collection serializer for responses.
+    ConnectionHookMetaData:
+      properties:
+        connection_type:
+          anyOf:
+          - type: string
+          - type: 'null'
+          title: Connection Type
+        hook_class_name:
+          anyOf:
+          - type: string
+          - type: 'null'
+          title: Hook Class Name
+        default_conn_name:
+          anyOf:
+          - type: string
+          - type: 'null'
+          title: Default Conn Name
+        hook_name:
+          type: string
+          title: Hook Name
+        standard_fields:
+          anyOf:
+          - $ref: '#/components/schemas/StandardHookFields'
+          - type: 'null'
+        extra_fields:
+          anyOf:
+          - type: object
+          - type: 'null'
+          title: Extra Fields
+      type: object
+      required:
+      - connection_type
+      - hook_class_name
+      - default_conn_name
+      - hook_name
+      - standard_fields
+      - extra_fields
+      title: ConnectionHookMetaData
+      description: 'Response model for Hook information == Connection type meta data.
+
+
+        It is used to transfer providers information loaded by providers_manager such
+        that
+
+        the API server/Web UI can use this data to render connection form UI.'
     ConnectionResponse:
       properties:
         connection_id:
@@ -9707,6 +9772,29 @@ components:
       - task_instance_states
       title: HistoricalMetricDataResponse
       description: Historical Metric Data serializer for responses.
+    HookFieldBehavior:
+      properties:
+        hidden:
+          type: boolean
+          title: Hidden
+          description: Flag if the form field should be hidden.
+          default: false
+        title:
+          anyOf:
+          - type: string
+          - type: 'null'
+          title: Title
+          description: Label / title for the field that should be displayed, if re-labelling
+            is needed. Use `None` to display standard title.
+        placeholder:
+          anyOf:
+          - type: string
+          - type: 'null'
+          title: Placeholder
+          description: Placeholder text that should be populated to the form.
+      type: object
+      title: HookFieldBehavior
+      description: A class to store the behavior of each standard field of a Hook.
     ImportErrorCollectionResponse:
       properties:
         import_errors:
@@ -10228,6 +10316,42 @@ components:
       - latest_scheduler_heartbeat
       title: SchedulerInfoResponse
       description: Scheduler info serializer for responses.
+    StandardHookFields:
+      properties:
+        description:
+          anyOf:
+          - $ref: '#/components/schemas/HookFieldBehavior'
+          - type: 'null'
+        url_schema:
+          anyOf:
+          - $ref: '#/components/schemas/HookFieldBehavior'
+          - type: 'null'
+        host:
+          anyOf:
+          - $ref: '#/components/schemas/HookFieldBehavior'
+          - type: 'null'
+        port:
+          anyOf:
+          - $ref: '#/components/schemas/HookFieldBehavior'
+          - type: 'null'
+        login:
+          anyOf:
+          - $ref: '#/components/schemas/HookFieldBehavior'
+          - type: 'null'
+        password:
+          anyOf:
+          - $ref: '#/components/schemas/HookFieldBehavior'
+          - type: 'null'
+      type: object
+      required:
+      - description
+      - url_schema
+      - host
+      - port
+      - login
+      - password
+      title: StandardHookFields
+      description: Standard fields of a Hook that a form will render.
     StructureDataResponse:
       properties:
         edges:
