+++ airflow/ui/openapi-gen/requests/schemas.gen.ts
@@ -1388,6 +1388,83 @@ export const $ConnectionCollectionResponse = {
   description: "Connection Collection serializer for responses.",
 } as const;
 
+export const $ConnectionHookMetaData = {
+  properties: {
+    connection_type: {
+      anyOf: [
+        {
+          type: "string",
+        },
+        {
+          type: "null",
+        },
+      ],
+      title: "Connection Type",
+    },
+    hook_class_name: {
+      anyOf: [
+        {
+          type: "string",
+        },
+        {
+          type: "null",
+        },
+      ],
+      title: "Hook Class Name",
+    },
+    default_conn_name: {
+      anyOf: [
+        {
+          type: "string",
+        },
+        {
+          type: "null",
+        },
+      ],
+      title: "Default Conn Name",
+    },
+    hook_name: {
+      type: "string",
+      title: "Hook Name",
+    },
+    standard_fields: {
+      anyOf: [
+        {
+          $ref: "#/components/schemas/StandardHookFields",
+        },
+        {
+          type: "null",
+        },
+      ],
+    },
+    extra_fields: {
+      anyOf: [
+        {
+          type: "object",
+        },
+        {
+          type: "null",
+        },
+      ],
+      title: "Extra Fields",
+    },
+  },
+  type: "object",
+  required: [
+    "connection_type",
+    "hook_class_name",
+    "default_conn_name",
+    "hook_name",
+    "standard_fields",
+    "extra_fields",
+  ],
+  title: "ConnectionHookMetaData",
+  description: `Response model for Hook information == Connection type meta data.
+
+It is used to transfer providers information loaded by providers_manager such that
+the API server/Web UI can use this data to render connection form UI.`,
+} as const;
+
 export const $ConnectionResponse = {
   properties: {
     connection_id: {
@@ -3837,6 +3914,45 @@ export const $HistoricalMetricDataResponse = {
   description: "Historical Metric Data serializer for responses.",
 } as const;
 
+export const $HookFieldBehavior = {
+  properties: {
+    hidden: {
+      type: "boolean",
+      title: "Hidden",
+      description: "Flag if the form field should be hidden.",
+      default: false,
+    },
+    title: {
+      anyOf: [
+        {
+          type: "string",
+        },
+        {
+          type: "null",
+        },
+      ],
+      title: "Title",
+      description:
+        "Label / title for the field that should be displayed, if re-labelling is needed. Use `None` to display standard title.",
+    },
+    placeholder: {
+      anyOf: [
+        {
+          type: "string",
+        },
+        {
+          type: "null",
+        },
+      ],
+      title: "Placeholder",
+      description: "Placeholder text that should be populated to the form.",
+    },
+  },
+  type: "object",
+  title: "HookFieldBehavior",
+  description: "A class to store the behavior of each standard field of a Hook.",
+} as const;
+
 export const $ImportErrorCollectionResponse = {
   properties: {
     import_errors: {
@@ -4586,6 +4702,75 @@ export const $SchedulerInfoResponse = {
   required: ["status", "latest_scheduler_heartbeat"],
   title: "SchedulerInfoResponse",
   description: "Scheduler info serializer for responses.",
+} as const;
+
+export const $StandardHookFields = {
+  properties: {
+    description: {
+      anyOf: [
+        {
+          $ref: "#/components/schemas/HookFieldBehavior",
+        },
+        {
+          type: "null",
+        },
+      ],
+    },
+    url_schema: {
+      anyOf: [
+        {
+          $ref: "#/components/schemas/HookFieldBehavior",
+        },
+        {
+          type: "null",
+        },
+      ],
+    },
+    host: {
+      anyOf: [
+        {
+          $ref: "#/components/schemas/HookFieldBehavior",
+        },
+        {
+          type: "null",
+        },
+      ],
+    },
+    port: {
+      anyOf: [
+        {
+          $ref: "#/components/schemas/HookFieldBehavior",
+        },
+        {
+          type: "null",
+        },
+      ],
+    },
+    login: {
+      anyOf: [
+        {
+          $ref: "#/components/schemas/HookFieldBehavior",
+        },
+        {
+          type: "null",
+        },
+      ],
+    },
+    password: {
+      anyOf: [
+        {
+          $ref: "#/components/schemas/HookFieldBehavior",
+        },
+        {
+          type: "null",
+        },
+      ],
+    },
+  },
+  type: "object",
+  required: ["description", "url_schema", "host", "port", "login", "password"],
+  title: "StandardHookFields",
+  description: "Standard fields of a Hook that a form will render.",
 } as const;
 
 export const $StructureDataResponse = {
